+++ src/facts_experiment_builder/core/experiment/__init__.py
@@ -1,5 +1,5 @@
 """Experiment-level types (in-memory representation of experiment-config.yaml)."""
 
-from facts_experiment_builder.core.experiment.facts_experiment import FactsExperiment
+from facts_experiment_builder.core.experiment.experiment import FactsExperiment
 
 __all__ = ["FactsExperiment"]
