+++ src/facts_experiment_builder/io/experiment_storage.py
@@ -1,8 +1,8 @@
-from facts_experiment_builder.core.experiment.paths import (
-    ExperimentPathContainer,
+from facts_experiment_builder.io.layout import (
+    ExperimentPaths,
 )
 
 
-def make_output_dir(experiment_paths: ExperimentPathContainer) -> None:
+def make_output_dir(experiment_paths: ExperimentPaths) -> None:
     output_dir = experiment_paths.output_dir
     output_dir.mkdir(parents=True)
