+++ src/facts_experiment_builder/io/module_registry.py
@@ -1,9 +1,11 @@
 from pathlib import Path
-from facts_experiment_builder.core.module.module_schema import ModuleSchema
-
 import subprocess
 from pydantic import ValidationError
 import yaml
+from typing import Protocol
+
+# ---------------------- Core imports ----------------------------
+from facts_experiment_builder.core.module.module_schema import ModuleSchema
 from facts_experiment_builder.io.exceptions import ModuleYamlNotFoundError
